Color coding reports.
Started using a fork of github.com/fatih/color package for color coding.
There was an issue in the way we were formating the cells and color
functions.

Added a new set of functions to the forked color package that supports
fmt.Formatter interface, and with that color coding is now lot easier
within reports.

With this implementation there is no need to attach a predicate to the
column or cells.